python_utils/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-utils"
version = "0.1.0"
description = "Shared Python utilities and database layer for transit-planner"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Transit Planner Team" }
]
dependencies = [
  "SQLAlchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "psycopg[binary]>=3.1,<4.0",
  "GeoAlchemy2>=0.15,<1.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.6,<1.0",
  "mypy>=1.10,<2.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["python_utils*"]

[tool.setuptools.package-data]
python_utils = ["py.typed"]

[tool.alembic]
script_location = "alembic"
prepend_sys_path = ["."]